Continuables

Continuables.Continuable

Continuables.azip

Continuables.chzip

Continuables.cont

Continuables.flatten

Continuables.groupbyreduce

Continuables.reduce!

Continuables.stoppable

Continuables.@ifsomething

Continuables.Continuable — Type

Continuable(func)

Assumes func to have a single argument, the continuation function (usually named cont).

Example

Continuable(function (cont)
  for i in 1:10
    cont(i)
  end
end)

Continuables.azip — Method

zipping continuables via intermediate array representation

CAUTION: loads everything into memory

Continuables.chzip — Method

zipping continuables via Channel

Continuables.cont — Method

cont is reserved function parameter name

Continuables.flatten — Method

for iterables of continuable use Continuables.chain(...)

Continuables.groupbyreduce — Function

group elements of continuable by by, aggregating immediately with op2/op1

Parameters

by: function of element to return the key for the grouping/dict continuable: will get grouped op2: f(accumulator, element) = newaccumulator op1: f(element) = initialaccumulator

Continuables.reduce! — Method

mutating version of reduce!

if no init is given op! is assumed to mutate a hidden state (equivalent to mere continuation) else init is the explicit state and will be passed to op! as first argument (the accumulator)

Continuables.stoppable — Function

contextmanager handling custom breakpoints with stop()

This is usually only used within creating a new continuable from a previous one Examples

@cont stop_at4(continuable) = stoppable(continuable) do x
  x == 4 && stop()
  cont(x)
end

Continuables.@ifsomething — Macro

IterTools.@ifsomething expr

If expr evaluates to nothing, equivalent to return nothing, otherwise the macro evaluates to the value of expr. Not exported, useful for implementing iterators.

julia> IterTools.@ifsomething iterate(1:2)
(1, 1)
julia> let elt, state = IterTools.@ifsomething iterate(1:2, 2); println("not reached"); end