Public API

Continuables.Continuable

Continuables.flatten

Continuables.groupbyreduce

Continuables.reduce!

Core

Continuables.Continuable — Type

Continuable(func)

Assumes func to have a single argument, the continuation function (usually named cont).

Example

Continuable(function (cont)
  for i in 1:10
    cont(i)
  end
end)

Missing docstring.

Missing docstring for @cont. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @Ref. Check Documenter's build log for details.

Conversions

Missing docstring.

Missing docstring for aschannel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ascontinuable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for i2c. Check Documenter's build log for details.

Missing docstring.

Missing docstring for @i2c. Check Documenter's build log for details.

Functionality

Missing docstring.

Missing docstring for singleton. Check Documenter's build log for details.

Missing docstring.

Missing docstring for repeated. Check Documenter's build log for details.

Missing docstring.

Missing docstring for iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reduce. Check Documenter's build log for details.

Continuables.reduce! — Function

mutating version of reduce!

if no init is given op! is assumed to mutate a hidden state (equivalent to mere continuation) else init is the explicit state and will be passed to op! as first argument (the accumulator)

Missing docstring.

Missing docstring for zip. Check Documenter's build log for details.

Missing docstring.

Missing docstring for product. Check Documenter's build log for details.

Missing docstring.

Missing docstring for chain. Check Documenter's build log for details.

Continuables.flatten — Function

for iterables of continuable use Continuables.chain(...)

Missing docstring.

Missing docstring for cycle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for foreach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for map. Check Documenter's build log for details.

Missing docstring.

Missing docstring for all. Check Documenter's build log for details.

Missing docstring.

Missing docstring for any. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sum. Check Documenter's build log for details.

Missing docstring.

Missing docstring for prod. Check Documenter's build log for details.

Missing docstring.

Missing docstring for take. Check Documenter's build log for details.

Missing docstring.

Missing docstring for takewhile. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drop. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dropwhile. Check Documenter's build log for details.

Missing docstring.

Missing docstring for partition. Check Documenter's build log for details.

Continuables.groupbyreduce — Function

group elements of continuable by by, aggregating immediately with op2/op1

Parameters

by: function of element to return the key for the grouping/dict continuable: will get grouped op2: f(accumulator, element) = newaccumulator op1: f(element) = initialaccumulator

Missing docstring.

Missing docstring for groupby, nth. Check Documenter's build log for details.

cont, @cont, Continuable, @Ref, stoppable, stop, singleton, repeated, iterate, aschannel, ascontinuable, i2c, @i2c, reduce, reduce!, zip, product, chain, flatten, cycle, foreach, map, all, any, sum, prod, take, takewhile, drop, dropwhile, partition, groupbyreduce, groupby, nth